+++ Basecode.WebApp/Startup.AutoMapper.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Basecode.Data.Dtos;
+using Basecode.Data.Dtos.HrEmployee;
 using Basecode.Data.Models;
 using Basecode.Data.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
