+++ Basecode.WebApp/Views/Admin/CreateHrAccount.cshtml
@@ -1,4 +1,5 @@
-﻿@using Basecode.Data.Dtos
+﻿@    using Basecode.Data.Dtos;
+    using Basecode.Data.Dtos.HrEmployee
 @model HREmployeeCreationDto
 @{
     ViewData["CreateHrAccount"] = "CREATE HR ACCOUNT";
