+++ Basecode.WebApp/Views/Admin/EditHrAccountView.cshtml
@@ -1,4 +1,5 @@
-﻿@using Basecode.Data.Dtos
+﻿@    using Basecode.Data.Dtos;
+    using Basecode.Data.Dtos.HrEmployee
 @model HREmployeeUpdationDto
 @{
     ViewData["EditHrAccount"] = "EDIT HR EMPLOYEE ACCOUNT INFORMATION";
