+++ Basecode.WebApp/Views/Login/Index.cshtml
@@ -1,4 +1,5 @@
-@using Basecode.Data.Dtos
+@    using Basecode.Data.Dtos;
+    using Basecode.Data.Dtos.HrEmployee
 @model HREmployeeUpdationDto
 @{
     ViewData["Title"] = "Login";
